Correct Function Return Moment and Output Statements, Code cleanup, and To-do comments added
See new comments on lines 88, 93, and 106.
These parts should be implemented.

Old resolved comments were removed.

Validating functions' returns were moved to end of function as implied by statement of the problem.

Output statements in each of the 3 validating functions should now be consistent with specific output requested in statement of the problem.